Updated the Prefabs and Camera Settings - Chris
I had to remove and remodify the camera controls for the player as it was throwing errors and it wasn't working as intended. Also made it so that when there is a collision with a wall or object, rather than seeing straight through it, the camera tries to smooth over the object and preserve the distance.